requirements.txt
langchain==0.1.10
mistralai==0.0.1
python-dotenv==1.0.0
textblob==0.17.1
pyyaml==6.0.1
